# app/Http/Controllers/Api/StudentController.php
<?php

namespace App\Http\Controllers\Api;

use App\Exceptions\ReservationCancellationException;
use App\Http\Controllers\Controller;
use App\Http\Requests\CancelReservationRequest;
use App\Http\Requests\ChangePasswordRequest;
use App\Http\Requests\UpdateStudentProfileRequest;
use App\Services\PaymentService;
use App\Services\ReservationService;
use App\Services\StudentService;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use OpenApi\Attributes as OA;

class StudentController extends Controller
{
    /**
     * Role ID used by students in the system.
     */
    private const STUDENT_ROLE_ID = 3;

    public function __construct(
        private readonly ReservationService $reservationService,
        private readonly StudentService $studentService,
        private readonly PaymentService $paymentService
    ) {}

    /**
     * Get the authenticated student's profile.
     */
    #[OA\Get(
        path: '/api/my-profile',
        summary: 'Get my profile',
        description: 'Returns the profile information of the authenticated student.',
        tags: ['Students'],
        security: [['sanctum' => []]],
        responses: [
            new OA\Response(
                response: 200,
                description: 'Profile retrieved successfully'
            ),
            new OA\Response(
                response: 401,
                description: 'Unauthenticated'
            ),
            new OA\Response(
                response: 403,
                description: 'Authenticated user is not a student'
            )
        ]
    )]
    public function myProfile(Request $request): JsonResponse
    {
        $student = $request->user();

        // Only students can access this resource.
        if ((int) $student->role_id !== self::STUDENT_ROLE_ID) {
            return response()->json([
                'success' => false,
                'message' => 'Only students can access this resource.',
            ], 403);
        }

        return response()->json([
            'success' => true,
            'message' => 'Profile retrieved successfully',
            'data' => $this->studentService->getMyProfile($student),
        ]);
    }

    /**
     * Update the authenticated student's profile.
     */
    #[OA\Put(
        path: '/api/my-profile',
        summary: 'Update my profile',
        description: 'Updates the profile information of the authenticated student.',
        tags: ['Students'],
        security: [['sanctum' => []]],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                properties: [
                    new OA\Property(
                        property: 'first_names',
                        type: 'string',
                        example: 'John'
                    ),
                    new OA\Property(
                        property: 'last_names',
                        type: 'string',
                        example: 'Smith'
                    ),
                    new OA\Property(
                        property: 'email',
                        type: 'string',
                        format: 'email',
                        example: 'john.smith@example.com'
                    ),
                    new OA\Property(
                        property: 'phone',
                        type: 'string',
                        example: '3001234567'
                    ),
                    new OA\Property(
                        property: 'photo',
                        type: 'string',
                        nullable: true,
                        example: 'photos/students/john.jpg'
                    )
                ]
            )
        ),
        responses: [
            new OA\Response(
                response: 200,
                description: 'Profile updated successfully'
            ),
            new OA\Response(
                response: 401,
                description: 'Unauthenticated'
            ),
            new OA\Response(
                response: 403,
                description: 'Authenticated user is not a student'
            ),
            new OA\Response(
                response: 422,
                description: 'Validation error'
            )
        ]
    )]
    public function updateMyProfile(
        UpdateStudentProfileRequest $request
    ): JsonResponse {
        $student = $request->user();

        $student = $this->studentService->updateMyProfile(
            $student,
            $request->validated()
        );

        return response()->json([
            'success' => true,
            'message' => 'Profile updated successfully',
            'data' => $student,
        ]);
    }

    /**
     * Change the authenticated student's password.
     */
    #[OA\Put(
        path: '/api/my-profile/change-password',
        summary: 'Change password',
        description: 'Changes the password of the authenticated student.',
        tags: ['Students'],
        security: [['sanctum' => []]],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                required: [
                    'current_password',
                    'new_password',
                    'new_password_confirmation'
                ],
                properties: [
                    new OA\Property(
                        property: 'current_password',
                        type: 'string',
                        format: 'password',
                        example: 'OldPassword123'
                    ),
                    new OA\Property(
                        property: 'new_password',
                        type: 'string',
                        format: 'password',
                        example: 'NewPassword123'
                    ),
                    new OA\Property(
                        property: 'new_password_confirmation',
                        type: 'string',
                        format: 'password',
                        example: 'NewPassword123'
                    )
                ]
            )
        ),
        responses: [
            new OA\Response(
                response: 200,
                description: 'Password changed successfully'
            ),
            new OA\Response(
                response: 401,
                description: 'Unauthenticated'
            ),
            new OA\Response(
                response: 403,
                description: 'Authenticated user is not a student'
            ),
            new OA\Response(
                response: 422,
                description: 'Validation error'
            )
        ]
    )]
    public function changePassword(
        ChangePasswordRequest $request
    ): JsonResponse {
        $student = $request->user();

        $this->studentService->changePassword(
            $student,
            $request->validated('new_password')
        );

        return response()->json([
            'success' => true,
            'message' => 'Password changed successfully',
        ]);
    }

    /**
     * Get reservations belonging to the authenticated student.
     */
    #[OA\Get(
        path: '/api/my-reservations',
        summary: 'Get my reservations',
        description: 'Returns the reservations of the authenticated student.',
        tags: ['Students'],
        security: [['sanctum' => []]],
        parameters: [
            new OA\Parameter(
                name: 'date',
                description: 'Filter reservations by date',
                in: 'query',
                required: false,
                schema: new OA\Schema(
                    type: 'string',
                    format: 'date'
                ),
                example: '2026-09-20'
            )
        ],
        responses: [
            new OA\Response(
                response: 200,
                description: 'Reservations retrieved successfully'
            ),
            new OA\Response(
                response: 401,
                description: 'Unauthenticated'
            ),
            new OA\Response(
                response: 403,
                description: 'Authenticated user is not a student'
            ),
            new OA\Response(
                response: 422,
                description: 'Invalid date'
            )
        ]
    )]
    public function myReservations(Request $request): JsonResponse
    {
        // Validate optional date filter.
        $request->validate([
            'date' => [
                'nullable',
                'date_format:Y-m-d',
            ],
        ]);

        $student = $request->user();

        // Only students can access their reservations.
        if ((int) $student->role_id !== self::STUDENT_ROLE_ID) {
            return response()->json([
                'success' => false,
                'message' => 'Only students can access their reservations.',
            ], 403);
        }

        $reservations = $this->reservationService->getMyReservations(
            $student,
            $request->query('date')
        );

        return response()->json([
            'success' => true,
            'message' => 'Reservations retrieved successfully',
            'data' => $reservations,
        ]);
    }

    /**
     * Cancel one of the authenticated student's reservations.
     */
    #[OA\Delete(
        path: '/api/reservations/{id}',
        summary: 'Cancel an authenticated student reservation',
        description: 'Cancels a confirmed reservation at least 24 hours before its start time. No refund is issued.',
        tags: ['Students'],
        security: [['sanctum' => []]],
        parameters: [
            new OA\Parameter(
                name: 'id',
                description: 'Reservation ID',
                in: 'path',
                required: true,
                schema: new OA\Schema(
                    type: 'integer',
                    example: 15
                )
            )
        ],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                required: ['cancellation_reason'],
                properties: [
                    new OA\Property(
                        property: 'cancellation_reason',
                        type: 'string',
                        minLength: 10,
                        maxLength: 500,
                        example: 'Tuve un imprevisto personal y no podré asistir.'
                    )
                ]
            )
        ),
        responses: [
            new OA\Response(
                response: 200,
                description: 'Reservation cancelled successfully'
            ),
            new OA\Response(
                response: 400,
                description: 'Business rule prevents cancellation'
            ),
            new OA\Response(
                response: 401,
                description: 'Unauthenticated'
            ),
            new OA\Response(
                response: 403,
                description: 'Not authorized'
            ),
            new OA\Response(
                response: 404,
                description: 'Reservation not found'
            ),
            new OA\Response(
                response: 422,
                description: 'Validation error'
            )
        ]
    )]
    public function cancelReservation(
        CancelReservationRequest $request,
        int $id
    ): JsonResponse {
        try {
            $this->reservationService->cancelReservation(
                $request->user(),
                $id,
                $request->validated('cancellation_reason')
            );

            return response()->json([
                'success' => true,
                'message' => 'Reservation cancelled successfully',
            ]);
        } catch (ReservationCancellationException $exception) {
            return response()->json([
                'success' => false,
                'message' => $exception->getMessage(),
            ], $exception->statusCode);
        }
    }

    /**
     * Get the authenticated student's completed payments.
     */
    #[OA\Get(
        path: '/api/my-payments',
        summary: 'Get my payments',
        description: 'Returns the payment history of the authenticated student.',
        tags: ['Students'],
        security: [['sanctum' => []]],
        parameters: [
            new OA\Parameter(
                name: 'date',
                description: 'Filter payments by reservation date',
                in: 'query',
                required: false,
                schema: new OA\Schema(
                    type: 'string',
                    format: 'date'
                ),
                example: '2026-09-20'
            )
        ],
        responses: [
            new OA\Response(
                response: 200,
                description: 'Payments retrieved successfully'
            ),
            new OA\Response(
                response: 401,
                description: 'Unauthenticated'
            ),
            new OA\Response(
                response: 403,
                description: 'Authenticated user is not a student'
            ),
            new OA\Response(
                response: 422,
                description: 'Invalid date'
            )
        ]
    )]
    public function myPayments(Request $request): JsonResponse
    {
        // Validate optional date filter.
        $request->validate([
            'date' => [
                'nullable',
                'date_format:Y-m-d',
            ],
        ]);

        $student = $request->user();

        // Only students can access their payments.
        if ((int) $student->role_id !== self::STUDENT_ROLE_ID) {
            return response()->json([
                'success' => false,
                'message' => 'Only students can access their payments.',
            ], 403);
        }

        try {
            $payments = $this->paymentService->getMyPayments(
                $student,
                $request->query('date')
            );

            return response()->json([
                'success' => true,
                'message' => 'Payments retrieved successfully',
                'data' => $payments,
            ]);
        } catch (\Throwable $e) {
            return response()->json([
                'success' => false,
                'message' => 'Error retrieving payments',
                'error' => $e->getMessage(),
            ], 500);
        }
    }
}