# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tutor Reservation API",
        "description": "RESTful API for tutor reservations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main server"
        }
    ],
    "paths": {
...
        "/api/my-profile": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get my profile",
                "description": "Returns the profile information of the authenticated student.",
                "operationId": "f63a4093ba3e452b443ee7b0bb826403",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a student"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update my profile",
                "description": "Updates the profile information of the authenticated student.",
                "operationId": "2406167f745cd0548ab0975943c4da94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_names": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_names": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.smith@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "3001234567"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "photos/students/john.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a student"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-profile/change-password": {
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Change password",
                "description": "Changes the password of the authenticated student.",
                "operationId": "85181866f3291d51f8b88561c07d3cf2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "OldPassword123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword123"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a student"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-reservations": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get my reservations",
                "description": "Returns the reservations of the authenticated student.",
                "operationId": "bedc6f45e89a04184456a0e1c81cb6e9",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter reservations by date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-09-20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservations retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a student"
                    },
                    "422": {
                        "description": "Invalid date"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Cancel an authenticated student reservation",
                "description": "Cancels a confirmed reservation at least 24 hours before its start time. No refund is issued.",
                "operationId": "556c78c34213ad52e7b4ab136aed8aa1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cancellation_reason"
                                ],
                                "properties": {
                                    "cancellation_reason": {
                                        "type": "string",
                                        "example": "Tuve un imprevisto personal y no podré asistir.",
                                        "maxLength": 500,
                                        "minLength": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation cancelled successfully"
                    },
                    "400": {
                        "description": "Business rule prevents cancellation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-payments": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get my payments",
                "description": "Returns the payment history of the authenticated student.",
                "operationId": "dc7c364d3cafe03b6dfdf7847dc0f494",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter payments by reservation date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-09-20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payments retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Authenticated user is not a student"
                    },
                    "422": {
                        "description": "Invalid date"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer token generated by Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}